Renamed layout into page-layout
