+++ src/pages/account/index.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 import { AccountOverview } from "@src/components/account-overview/account-overview";
 import { Container } from "@src/components/containers/container";
-import { Layout } from "@src/components/layout";
+import { Layout } from "@src/components/page-layout";
 import { deviceBreakpoints } from "@src/design-system/theme";
 import { withAuth } from "@src/middlewares/with-auth";
 import { withLogger } from "@src/middlewares/with-logger";
