+++ src/pages/account/locale.tsx
@@ -1,8 +1,8 @@
 import type { GetServerSideProps } from "next";
 import React from "react";
 
-import { Layout } from "@src/components/layout";
 import { Locale } from "@src/components/locale/locale";
+import { Layout } from "@src/components/page-layout";
 import { withAuth } from "@src/middlewares/with-auth";
 import { withLogger } from "@src/middlewares/with-logger";
 import { withSentry } from "@src/middlewares/with-sentry";
