+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import { getProviderName } from "@lib/queries/get-provider-name";
 import { NoDataReturned, NoProviderNameFound } from "@src/client-errors";
 import { Home } from "@src/components/home/home";
-import { Main } from "@src/components/layout";
+import { Main } from "@src/components/page-layout";
 import { oauth2Client } from "@src/config";
 import { GraphqlErrors } from "@src/errors";
 import { withLogger } from "@src/middlewares/with-logger";
